+++ handpose/loss.py
@@ -296,7 +296,16 @@ def kpt_conf_loss(k_conf_truth, k_conf_pred, obj_conf, lambda_kpt_conf=1):
     return lambda_kpt_conf * loss
 
 def loss_fn(truth, prediction, lambda_coord=5, lambda_noobj=0.5, epsilon=1e-6, lambda_kpt=0.5, lambda_kpt_conf=0.5):
-    r"""Computes loss.
+    r"""Computes loss and returns a dictionary of all the losses.
+
+    The losses returned are:
+    
+    - Total Loss
+    - Box Loss
+    - Class Loss
+    - Keypoint Loss
+    - Keypoint Confidence Loss
+
     
     Parameters
     ----------
@@ -317,7 +326,7 @@ def loss_fn(truth, prediction, lambda_coord=5, lambda_noobj=0.5, epsilon=1e-6, l
 
     Returns
     -------
-    torch.tensor
+    dict
 
     Examples
     --------
@@ -392,4 +401,12 @@ def loss_fn(truth, prediction, lambda_coord=5, lambda_noobj=0.5, epsilon=1e-6, l
     # Adding all the losses
     loss = L_box + L_conf + L_class + L_kpt + L_kpt_conf
 
-    return loss
+    all_losses = {'total_loss': loss, 
+                  'box_loss': L_box,
+                  'conf_loss': L_conf,
+                  'class_loss': L_class,
+                  'kpt_loss': L_kpt,
+                  'kpt_conf_loss': L_kpt_conf 
+                 }
+
+    return all_losses